src/main/java/com/em/demo/java/springwebapp/service/UserService.java
package com.em.demo.java.springwebapp.service;

import com.em.demo.java.springwebapp.model.entity.User;
import com.em.demo.java.springwebapp.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class UserService {

    @Autowired
    private UserRepository userRepository;

    public List<User> getAll() {
        return userRepository.findAll();
    }

    public User save(User user) {
        return userRepository.save(user);
    }
}
